v1.0.1: fix the way encryption keys were managed
previously, only one key was supported per stream manifest;
and the code that managed the key's filename wasn't quite right.

now, each manifest can utilize as many encryption keys as it wants.
each key is assigned a sequential name,
and this new filename is reflected correctly in the local manifest.